easy_ocr_app/requirements.txt
torch==2.6.0
torchvision>=0.5
opencv-python-headless==4.11.0.86
scipy==1.15.2
numpy==2.2.4
Pillow==11.1.0
scikit-image==0.25.2
python-bidi==0.6.6
PyYAML==6.0.2
Shapely==2.0.7
pyclipper== 1.3.0.post6
ninja==1.11.1.4
natsort==8.4.0
python-bidi==0.6.6
easyocr==1.7.2
pillow==11.1.0
Flask==3.1.0

curl -X POST \
  http://localhost:5002/process_image \
  -H "Content-Type: application/json" \
  -d "{\"image\": \"$(base64 -i /Users/chervonikov_alexey/Desktop/invalid_images/IMG_8337.PNG | tr -d '\n')\"}"
